+++ tests/conftest.py
@@ -10,7 +10,6 @@ sys.path.insert(0, str(pathlib.PurePath(__file__).parent.parent / 'src'))
 from galaxy.api.errors import UnknownError
 from plugin import HumbleBundlePlugin
 from settings import Settings
-from model.game import Key, Subproduct, TroveGame
 
 
 class AsyncMock(MagicMock):
@@ -104,22 +103,3 @@ def get_troves(get_data):
             troves.extend(get_data(f'troves_{i + 1}.json'))
         return troves
     return fn
-
-
-@pytest.fixture
-def get_torchlight_trove(get_troves):
-    troves_data = get_troves(from_index=0)
-    for i in troves_data:
-        if i['machine_name'] == 'torchlight_trove':
-            trove_torchlight = i
-    return trove_torchlight, TroveGame(trove_torchlight)
-
-
-@pytest.fixture
-def get_torchlight(orders_keys):
-    for i in orders_keys:
-        if i['product']['machine_name'] == 'torchlight_storefront':
-            torchlight_order = i
-    drm_free = Subproduct(torchlight_order['subproducts'][0])
-    key = Key(torchlight_order['tpkd_dict']['all_tpks'][0])
-    return torchlight_order, drm_free, key
